+++ Package.swift
@@ -7,7 +7,8 @@ var dependencies: [Package.Dependency] = [
     .package(url: "https://github.com/tadija/AEXML", from: "4.0.0"),
     .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
     .package(url: "https://github.com/kateinoigakukun/swift-indexstore", from: "0.0.0"),
-    .package(url: "https://github.com/peripheryapp/swift-syntax", .exact("1.0.1"))
+    .package(url: "https://github.com/peripheryapp/swift-syntax", .exact("1.0.1")),
+    .package(url: "https://github.com/ileitch/swift-filename-matcher", from: "0.0.0")
 ]
 
 #if os(macOS)
@@ -23,7 +24,8 @@ dependencies.append(
 var frontendDependencies: [PackageDescription.Target.Dependency] = [
     .target(name: "Shared"),
     .target(name: "PeripheryKit"),
-    .product(name: "ArgumentParser", package: "swift-argument-parser")
+    .product(name: "ArgumentParser", package: "swift-argument-parser"),
+    .product(name: "FilenameMatcher", package: "swift-filename-matcher")
 ]
 
 #if os(macOS)
@@ -43,14 +45,16 @@ var targets: [PackageDescription.Target] = [
             .product(name: "AEXML", package: "AEXML"),
             .product(name: "SwiftSyntax", package: "swift-syntax"),
             .product(name: "SwiftSyntaxParser", package: "swift-syntax"),
-            .product(name: "SwiftIndexStore", package: "swift-indexstore")
+            .product(name: "SwiftIndexStore", package: "swift-indexstore"),
+            .product(name: "FilenameMatcher", package: "swift-filename-matcher")
         ]
     ),
     .target(
         name: "Shared",
         dependencies: [
             .product(name: "Yams", package: "Yams"),
-            .product(name: "SystemPackage", package: "swift-system")
+            .product(name: "SystemPackage", package: "swift-system"),
+            .product(name: "FilenameMatcher", package: "swift-filename-matcher")
         ]
     ),
     .target(